Refactor code for composition
